feat: Enhance error handling for student operations by returning 404 responses when a student is not found in index.php and repository.php

+++ jiuflix-php/repositories/repository.php
@@ -44,6 +44,17 @@ class AlunosRepository
     {
         $pdo = connectionDatabase();
 
+        $sqlId = "SELECT * FROM aluno WHERE id = :id";
+        $stmt = $pdo->prepare($sqlId);
+        $stmt->bindValue(':id', $id);
+        $stmt->execute();
+
+        $result = $stmt->fetch(PDO::FETCH_OBJ);
+
+        if (!$result) {
+            return null;
+        }
+
         $sql = "UPDATE aluno SET name = :name, type_graduation = :type_graduation WHERE id = :id";
 
         $stmt = $pdo->prepare($sql);
@@ -51,13 +62,6 @@ class AlunosRepository
         $stmt->bindValue(':type_graduation', $typeGraduation, PDO::PARAM_STR);
         $stmt->bindValue(':id', $id, PDO::PARAM_INT);
         $stmt->execute();
-
-        $sqlId = "SELECT * FROM aluno WHERE id = :id";
-        $stmt = $pdo->prepare($sqlId);
-        $stmt->bindValue(':id', $id);
-        $stmt->execute();
-
-        $result = $stmt->fetch(PDO::FETCH_OBJ);
 
         return $result;
     }
@@ -78,11 +82,25 @@ class AlunosRepository
     public static function deleted($id)
     {
         $pdo = connectionDatabase();
+
+
+        $sqlId = "SELECT * FROM aluno WHERE id = :id";
+        $stmt = $pdo->prepare($sqlId);
+        $stmt->bindValue(':id', $id);
+        $stmt->execute();
+
+        $result = $stmt->fetch(PDO::FETCH_OBJ);
+
+        if (!$result) {
+            return null;
+        }
     
         $sql = "DELETE FROM aluno WHERE id = ?";
     
         $stmt = $pdo->prepare($sql);
         $stmt->bindValue(1, $id, PDO::PARAM_INT);
         $stmt->execute();
+
+        return $result;
     }
 }
